pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image_utils"
version = "0.0.1"
authors = [
  { name="Alexander Swerdlow", email="aswerdlow1@gmail.com" },
]
description = "A small example package"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "opencv-python>=4.5",
    "numpy>=1.22",
    "torch>=1.12",
    "torchvision>=0.13",
    "pillow>=9.0.0",
    "einops>=0.6.0",
    "StrEnum>=0.4.9",
    ""
]

[project.optional-dependencies]
VIDEO = ["av>=10.0.0"]
PROC = ["joblib>=1.2.0"]

[project.urls]
Homepage = "https://github.com/alexanderswerdlow/image_utils"